clean up

+++ data/scripts/download-osm.sh
@@ -9,6 +9,8 @@ mkdir -p $outputDir
 
 for city in "$@"
 do
+  outFile=$outputDir/$city.osm
+  [[ -e $outFile  ]] && continue
   echo "=== Downloading $city ==="
   wget -O $outputDir/$city.osm $query\&bbox=$(jq -r ".\"${city/_/ }\" | join(\",\")" city-bbox-index.json)
 done
